Removed TokenValidationError since it is no longer used
